Move useBlockBindingsUtils
