+++ packages/block-editor/src/components/block-bindings/index.js
@@ -3,3 +3,4 @@
  */
 export { default as BlockBindingsAttributeControl } from './attribute-control';
 export { default as BlockBindingsSourceFieldsList } from './source-fields-list';
+export { default as useBlockBindingsUtils } from './use-block-bindings-utils';
