+++ packages/block-editor/src/components/index.js
@@ -14,6 +14,7 @@ export {
 export {
 	BlockBindingsAttributeControl,
 	BlockBindingsSourceFieldsList,
+	useBlockBindingsUtils,
 } from './block-bindings';
 export { default as __experimentalBlockFullHeightAligmentControl } from './block-full-height-alignment-control';
 export { default as __experimentalBlockAlignmentMatrixControl } from './block-alignment-matrix-control';
