+++ packages/block-editor/src/utils/index.js
@@ -1,3 +1,2 @@
 export { default as transformStyles } from './transform-styles';
 export { default as getPxFromCssUnit } from './get-px-from-css-unit';
-export { useBlockBindingsUtils } from './block-bindings';
